packages/coding_problem_generator/pyproject.toml
[tool.poetry]
name = "coding_problem_generator"
version = "0.0.1"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
langchain-core = ">=0.1.5"
langchain-openai = ">=0.0.1"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.4"
fastapi = "^0.104.0"
sse-starlette = "^1.6.5"

[tool.langserve]
export_module = "coding_problem_generator"
export_attr = "chain"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
